I concede
There is just too many issues for me to want to continue messing around here. This is as far as I got. I tried running XDM in a VM on macOS using the most recent release, and it doesn't seem to want to pull plugins down. Just going to move on and do my own thing. Will look back to XDM for inspiration along the way.